refactor(ui): extract class that validate user input
